Adds ingress routes for applications
When relating to an nginx-ingress-integrator charm, a Kubernetes
Ingress object is also created containing our desired information.

By default, the Kubernetes Ingress object is created with the path ``/``,
which means that we won't be able to use the same hostname for ingress.
Additionally, it has the annotation ``nginx.ingress.kubernetes.io/rewrite-target: /``,
which means that if we would update the path to ``/legend-engine``, when
accessing Legend Engine through Ingress, that ``/legend-engine`` will be
replaced by ``/``.

Using "path-routes" will allow us to configure multiple applications to
have the same external hostnames, and the applications can be accessed
through these ingress routes.

For example, we can configure FINOS Legend Studio, Engine, and SDLC to
have the same external-hostname, and they would be reached via:

http://finos-legend/api -> Legend SDLC
http://finos-legend/engine -> Legend Engine
http://finos-legend/ -> Legend Studio

In order to use multiple routes as exemplified above, we would also have
to remove the ``nginx.ingress.kubernetes.io/rewrite-target`` annotation.
This can be done by using the ``rewrite-enabled = False`` config option
in the nginx-ingress-integrator charm.

Adds _get_ingress_routes method, which is to be implemented by the
LEgend SDLC, Engine, and Studio charms.